akademik-takvim/requirements.txt
django
djangorestframework
icecream
djangorestframework-simplejwt
pyjwt
django-cors-headers